python_engine/requirements.txt
numpy
scipy
dash
plotly
pyserial
pygame
h5py